pyproject.toml
[project]
name = "uweather-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-rest-auth>=7.0.0",
    "django>=5.1.3",
    "djangorestframework-simplejwt>=5.3.1",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
    "psycopg[binary]>=3.2.3",
    "pydantic-settings>=2.6.1",
    "pytest-django>=4.9.0",
    "ruff>=0.7.4",
]
